[Fix] 투표 상세 조회 DTO에 프로퍼티 추가하여 디코딩 문제 해결
